requirements.txt
python-decouple
mercantile
opencv-python
rasterio
requests
tqdm
